Projeto completado com sucesso

+++ src/business/LikeDislikeBusiness.ts
@@ -1,6 +1,6 @@
 import { LikeDislikeDataBase } from "../database/LikeDislikeDatabase";
 import { PostDatabase } from "../database/PostDatabase";
-import { UpdateLikeDislikeInputDTO } from "../dtos/dto-likedis/updateLikeDislike.dto";
+import { UpdateLikeDislikeInputDTO, UpdateLikeDislikeOutputDTO } from "../dtos/dto-likedis/updateLikeDislike.dto";
 import { BadRequest } from "../errors/BadRequest";
 import { POST_LIKE } from "../models/LikeDislike";
 import { Post } from "../models/Post";
@@ -13,14 +13,14 @@ export class LikeDislikeBusiness {
         private tokenManager: TokenManager
     ) { }
 
-    public updateLikeDislike = async (input: UpdateLikeDislikeInputDTO): Promise<void> => {
+    public updateLikeDislike = async (input: UpdateLikeDislikeInputDTO): Promise<UpdateLikeDislikeOutputDTO | undefined> => {
         const {
             id,
             token,
             like
         } = input
 
-        const payload = this.tokenManager.getPayload(token)
+        const payload = await this.tokenManager.getPayload(token)
 
         if (!payload) {
             throw new BadRequest("Faça o login para interagir nos posts")
@@ -32,6 +32,9 @@ export class LikeDislikeBusiness {
             throw new BadRequest("Post não encontrado")
         }
 
+        if (postDB.creator_id === payload.id) {
+            throw new BadRequest("Você não pode curtir ou discurtir o seu próprio post")
+        }
         const post = new Post(
             postDB.id,
             postDB.content,
@@ -43,7 +46,7 @@ export class LikeDislikeBusiness {
             postDB.creator_name
         )
 
-        const likeDislikeExist = await this.likeDislikeDatabase.verifyExistenceOfLikeDislike(postDB)
+        const likeDislikeExist = await this.likeDislikeDatabase.verifyExistenceOfLikeDislike(payload.id , id)
 
         const likeToSQLite = like ? 1 : 0
 
@@ -53,31 +56,55 @@ export class LikeDislikeBusiness {
             like: likeToSQLite
         }
 
-        if (likeDislikeExist === POST_LIKE.LIKED) {
-            if (like) {
-                await this.likeDislikeDatabase.removeLikeDislike(likeDislikeDB)
-                post.removeLike()
-            } else {
-                await this.likeDislikeDatabase.updateLikeDislike(likeDislikeDB)
-                post.removeLike()
-                post.addDislike()
+        let message
+        
+        if(likeDislikeExist) {
+            if (likeDislikeExist === POST_LIKE.LIKED) {
+                if (like) {
+                    await this.likeDislikeDatabase.removeLikeDislike(likeDislikeDB)
+                    post.removeLike()
+                    message = "Você removeu seu like"
+                } else {
+                    await this.likeDislikeDatabase.updateLikeDislike(likeDislikeDB)
+                    post.removeLike()
+                    post.addDislike()
+                    message = "Você deu like"
+                }
+            } else if (likeDislikeExist === POST_LIKE.DISLIKED) {
+                if (!like) {
+                    await this.likeDislikeDatabase.removeLikeDislike(likeDislikeDB)
+                    post.removeDislike()
+                    message = "Você removeu seu dislike"
+                } else {
+                    await this.likeDislikeDatabase.updateLikeDislike(likeDislikeDB)
+                    post.removeDislike()
+                    post.addLike()
+                    message = "Você deu dislike"
+                }
             }
-        } else if (likeDislikeExist === POST_LIKE.DISLIKED) {
-            if (!like) {
-                await this.likeDislikeDatabase.removeLikeDislike(likeDislikeDB)
-                post.removeDislike()
-            } else {
-                await this.likeDislikeDatabase.updateLikeDislike(likeDislikeDB)
-                post.removeDislike()
-                post.addLike()
-            }
-        } else {
+        }else {
             await this.likeDislikeDatabase.newLikeDislike(likeDislikeDB)
-            like ? post.addLike() : post.addDislike()
+
+            const liked = () => {
+                post.addLike()
+                message = "Você deu like"
+            }
+            const disliked = () => {
+                post.addDislike()
+                message = "Você deu dislike"
+            }
+
+            like ? liked() : disliked()
         }
 
         const updatedPost = post.postToDB()
 
         await this.postDatabase.editPost(updatedPost)
+
+        const output = {
+            message
+        }
+
+        return output
     }
 }
